Fix inability to select default clock color option
Default clock color option could not be selected due to the fact that
the default clock color id was null. Therefore, when it is selected, the
overriding clock color id is set to null. This is confounding because
the overriding value being null could signal either that the default
clock color was selected, or no overriding clock color was selected, and
in the current logic, we assume the latter, and therefore provide the
system selected color id instead. Give default clock color option an id
to fix the problem.

Flag: com.android.systemui.shared.new_customization_picker_ui
Test: Manually verified
Test: ClockPickerViewModelTest passes
Bug: 369114305

Change-Id: I425da6159b338f80e6ded7fdb7a9240b062dd456