Send messages to Launcher preview when previewing color
When the user selects a color in the color picker, generate a list of
color ids and another list of their corresponding color overrides, and
send messages to the Launcher preview to update accordingly.

Color generation was previously done in the Launcher
MaterialColorsGenerator, but now moved to the picker.

Flag: com.android.systemui.shared.new_customization_picker_ui
Test: manually verified that colors are previewed and not applied
Bug: 350718581
Change-Id: I839afd886e23df4d505517a5b7a60108855fa01c