Clean up color options
Small refactor to facilitate upcoming color picker changes. Previously,
when creating color options from color seeds, the color seeds are
converted to strings and stored in a map of strings to mimic a json
package. In addition, default colors are given an empty string to
indicate that they are using the primary wallpaper color seed. Now that
we have use cases to log color seeds, and to preview UI colors based on
a selected color seed, it makes sense to explicitly store the color seed
integer in the color option.

Flag: EXEMPT small refactor
Test: Manually verified color picker options appear and apply correctly
Test: Manually verified that color option resets correctly
Test: unit tests
Bug: 350718581
Change-Id: Ib43094656dad804cf7f5edb3f5f3dc7506f0a510