Refactor color picker to better track selected option (1/2)
The color picker workflow used to involve creating a list of
ColorOptions, converting each option to a ColorOptionModel in the
repository, and then converting each option to an OptionItemViewModel in
the view model. The selected state of each color option used to be kept
individually in each ColorOptionModel. Create a new flagged color picker
workflow, and centralize the selected state by keeping a single
selectedColorOption variable in ColorPickerRepository2. Also remove the
usage of ColorOptionModel in the new flow completely since this
abstraction layer is no longer needed. This should simplify the new flow
and reduce bugs, and will simplify upcoming work.

Flag: com.android.systemui.shared.new_customization_picker_ui
Test: manually verified & unit tests
Bug: 350718581
Change-Id: Ibec699fc45e4ebd8b3fa37dc6c4cb23e099ff2f9