Fix color option update animation
When changing tabs in color picker, the recycler view color options are
replaced with a new list, with item animation. This animation is
proceeded by an undesirable shift in the options, due to the fact that
item decoration is set on the items, and this decoration is removed
before animating. Since we only need edge item spacing, we can remove
the item decoration and set padding on the recycler view instead, and
keep clip children as false. This achieves the same layout without
animation jank.

Flag: com.android.systemui.shared.new_customization_picker_ui
Bug: 376295026
Test: manually verified, see bug
Change-Id: I2efa16ca81a5126ff7b7a9d8398a24948b1842f2