Fix discard changes show up if apply the shortcut changes
Fixed by introducing an optimistic update of the quick affordances. This
insure that the onApply flow is updated before we return to the main
screen, while acutally updating the system quick affordances is async
and can not guarantee if the update can happen before or after the
turning back to the main screen.

Test: Manually tested that the apply button never triggers the dialog
Fixes: 388728062
Flag: com.android.systemui.shared.new_customization_picker_ui
Change-Id: Ie6e8451d0edc5522f2ead2a5fff89aaf272b4a75