Fix unresponsive color options in color section
On click listener is set on the outer color option item view, however,
the inner color option view was set to be clickable and focusable as
part of a Talkback fix, which caused regular user taps to not trigger
the on click listener. Set clickable and focusable on the item view
instead, and made sure Talkback still works as intended.

Flag: EXEMPT bug fix
Bug: 390539980
Test: manually verified the selected option is announced in Talkback
Test: manually verified options are selectable in and out of Talkback
Change-Id: I011757f4c10a12aada3768477eea064d60dabd93